split actor-sheet.js to allow multiple actor types
